diff-index.txt: update raw output format in examples
The two examples in the doc for 'git diff-index' were not updated when
the raw output format was changed in 986ded8b56 ([PATCH] The diff-raw
format updates., 2005-05-21) (first example) and in 4330d8ee2e ([PATCH]
diff-raw format update take #2., 2005-05-23) and b8611cf248 (diff:
diff_aligned_abbrev: remove ellipsis after abbreviated SHA-1 value,
2017-12-03) (second example).

Update the output, inventing some characters to complete the source
hash in the second example. Also correct the destination mode in the
second example, which was wrongly '100664' since the addition of the
example in 612c4016dc (Reference documentation for the core git
commands., 2005-05-05).

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>